Slit files
